backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
email-validator==2.1.0.post1